Servicios y Impl en planta y ejemplar codigo base

+++ tarea3dwes/src/main/java/com/pelayora/tarea3dwes/repositorios/PlantaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.pelayora.tarea3dwes.modelo.Planta;
 
 @Repository
-public interface PlantaRepository extends JpaRepository<Planta, Integer> {
+public interface PlantaRepository extends JpaRepository<Planta, Long> {
 
 }
